Opportunity_Discovery_Workflow/requirements.txt
# Opportunity Discovery Workflow - Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.26.0

# Database
aiosqlite>=0.19.0

# Agno Framework (AI Agents)
agno>=2.2.0

# OpenAI
openai>=1.12.0

# Data Processing
pandas>=2.0.0

# Utilities
python-multipart>=0.0.6
